IGNITE-23554 ScanQuery code cleanup
